Combined TeamsHomeFragment and TeamSetupFragment
