+++ app/src/main/res/layout/fragment_team_setup.xml
@@ -8,7 +8,8 @@
 
     <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:layout_marginTop="70dp">
         <GridLayout
             android:id="@+id/team_grid_layout"
             android:layout_width="match_parent"
